+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - ADD: Support for Status reports.
 - FIX: Create PoI from APRS data received over FlexPacket.
+- FIX: Error during cleanup old APRS data.
+- ADD: Now we can configure how old the APRS Data could be before we clean them up.
 
 ## v0.2.0